IOT-1583: Fixing /W3 warnings in /resource directory - part 1.
Fixing all /W4s inside the IPCA directory and a few minor
once inside service/.

Reviewed-on: https://gerrit.iotivity.org/gerrit/19555
Tested-by: jenkins-iotivity <jenkins@iotivity.org>

+++ resource/IPCA/samples/ElevatorClient/ElevatorClient.cpp
@@ -225,7 +225,7 @@ void DiscoverDevicesCallback(void* context,
     std::string deviceId = deviceInfo->deviceId;
     std::string deviceName = deviceInfo->deviceName;
     std::vector<std::string> deviceUris;
-    for (int i = 0; i < deviceInfo->deviceUriCount; i++)
+    for (size_t i = 0; i < deviceInfo->deviceUriCount; i++)
     {
         deviceUris.push_back(deviceInfo->deviceUris[i]);
     }
